[issue-19-domain] fix: domain test code to coroutine
