--- api/domain/src/test/kotlin/linkpool-v0/jobgroup/model/JobGroupTest.kt
@@ -1,8 +1,0 @@
-package linkpool.jobgroup.model
-
-import io.kotest.core.spec.style.BehaviorSpec
-import org.junit.jupiter.api.Assertions.*
-
-class JobGroupTest: BehaviorSpec({
-
-})
