--- api/domain/src/test/kotlin/linkpool-v0/link/model/LinkTest.kt
@@ -1,8 +1,0 @@
-package linkpool.link.model
-
-import io.kotest.core.spec.style.BehaviorSpec
-import org.junit.jupiter.api.Assertions.*
-
-class LinkTest: BehaviorSpec({
-
-})
