--- api/domain/src/test/kotlin/linkpool-v0/link/service/LinkServiceTest.kt
@@ -1,8 +1,0 @@
-package linkpool.link.service
-
-import io.kotest.core.spec.style.BehaviorSpec
-import org.junit.jupiter.api.Assertions.*
-
-class LinkServiceTest: BehaviorSpec({
-
-})
