--- api/domain/src/test/kotlin/linkpool-v0/query/service/UserJobGroupServiceTest.kt
@@ -1,8 +1,0 @@
-package linkpool.query.service
-
-import io.kotest.core.spec.style.BehaviorSpec
-import org.junit.jupiter.api.Assertions.*
-
-class UserJobGroupServiceTest: BehaviorSpec({
-
-})
